API

SparseGaussianProcesses.CircularSquaredExponentialKernel — Type

CircularSquaredExponentialKernel

A squared exponential kernel on the circle, defined on the circle or, if $d > 1$, the $d$-dimensional torus. The expression is given by

$k(x,x') = \sum_{n \in 2\pi\mathbb{Z}^d} \sigma^2\exp\left( -\left\lVert\frac{x - x' + n}{\kappa} \right\rVert^2 \right)$

parameterized by log-variance $\ln(\sigma^2)$ (trainable by default) and log-length-scales $\ln(\kappa)$ (trainable by default) applied element-wise to each dimension. The sum is computed to a specified truncation level.

SparseGaussianProcesses.CircularSquaredExponentialKernel — Method

(k::CircularSquaredExponentialKernel)(x1::AbstractMatrix, 
                                      x2::AbstractMatrix)

Computes the kernel matrix for the given circular squared exponential kernel.

SparseGaussianProcesses.CircularSquaredExponentialKernel — Method

CircularSquaredExponentialKernel(dim::Int)

Creates a squared exponential kernel defined on the circle, or more generally a torus of dimension dim.

SparseGaussianProcesses.EuclideanRandomFeatures — Type

EuclideanRandomFeatures

A set of Euclidean random features, parameterized by frequency and phase, together with a set of associated basis weights.

SparseGaussianProcesses.EuclideanRandomFeatures — Method

(self::EuclideanRandomFeatures)(x::AbstractArray{<:Any,3},w::AbstractMatrix, 
                                k::GradientKernel{<:EuclideanKernel})

Evaluate $(\nabla g)(x)$ where $g$ is a Gaussian process with kernel $k$, $\nabla$ is the gradient inter-domain operator, and $x$ is the batched data, using the random features.

SparseGaussianProcesses.EuclideanRandomFeatures — Method

(self::EuclideanRandomFeatures)(x::AbstractMatrix, w::AbstractMatrix, 
                                k::GradientKernel{<:EuclideanKernel})

Evaluate $(\nabla g)(x)$ where $g$ is a Gaussian process with kernel $k$, $\nabla$ is the gradient inter-domain operator, and $x$ is the data, using the random features.

SparseGaussianProcesses.EuclideanRandomFeatures — Method

(self::EuclideanRandomFeatures)(x::AbstractMatrix, w::AbstractMatrix, 
                                k::EuclideanKernel)

Evaluate the $f(x)$ where $f$ is a Gaussian process with kernel $k$, and $x$ is the data, using the random features.

SparseGaussianProcesses.EuclideanRandomFeatures — Method

EuclideanRandomFeatures(k::EuclideanKernel, num_features::Int)

Create a set of Euclidean random features with eigenvalues given by the spectral distribution given by $k$.

SparseGaussianProcesses.GradientKernel — Method

(k::GradientKernel{<:SquaredExponentialKernel})(x1::AbstractMatrix, 
                                                x2::AbstractMatrix)

Computes the kernel matrix for the given gradient squared exponential kernel.

SparseGaussianProcesses.GradientOperator — Type

GradientOperator

The gradient inter-domain operator $\mathcal{A}f = \nabla f$.

SparseGaussianProcesses.IdentityOperator — Type

IdentityOperator

The identity inter-domain operator $\mathcal{A}f = f$.

SparseGaussianProcesses.MarginalInducingPoints — Type

MarginalInducingPoints

A set of inducing points representing the marginal value of the Gaussian process $u = (\mathcal{B} g)(z)$.

SparseGaussianProcesses.MarginalInducingPoints — Method

(self::MarginalInducingPoints)(z::AbstractMatrix, k::Kernel)

Sets the inducing locations of $u$ to $z$, inducing mean to zero, and inducing covariance to $k(z,z)$.

SparseGaussianProcesses.MarginalInducingPoints — Method

MarginalInducingPoints(k::CovarianceKernel, num_inducing::Integer)

Creates a set of marginal inducing points with covariance kernel $k$ of $u$.

SparseGaussianProcesses.MarginalInducingPoints — Method

(self::MarginalInducingPoints)()

Assembles the inducing covariance into upper-triangular form, with diagonal values exponentiated to ensure they are positive. Returns inducing locations, inducing mean, jitter covariance, and inducing covariance.

SparseGaussianProcesses.NormalHyperprior — Type

NormalHyperprior

An isotropic multivariate normal hyperprior, parameterized by mean and element-wise standard deviation.

SparseGaussianProcesses.NormalHyperprior — Method

(hp::NormalHyperprior)(x::AbstractVector)

Evaluates the normal hyperprior at $x$.

SparseGaussianProcesses.SparseGaussianProcess — Type

SparseGaussianProcess

A sparse Gaussian process, with kernel kernel. It supports models of the form

$(f | u)(\cdot) = (\mathcal{A} g)(\cdot) + K_{(\cdot)z} (K_{zz} + \Lambda)^{-1} (u - (\mathcal{B} g)(z) - \epsilon)$

where $g \sim GP(0, k)$, $u \sim N(\mu, \Sigma)$, $\epsilon \sim N(0, \Lambda)$, and $\mathcal{A}$, $\mathcal{B}$ are inter-domain operators.

Fields

kernel: the kernel $k$ of $g$.

observation_operator: the observation operator $\mathcal{A}$.

inter_domain_operator: the inter-domain operator $\mathcal{B}$.

prior_basis: the basis and weights used for efficiently sampling the prior.

inducing_points: the basis and weights used for sampling the data-dependent portion of the GP.

log_error: the error variance (log-scale, trainable by default).

hyperprior: the hyperprior used for the log error term.

SparseGaussianProcesses.SparseGaussianProcess — Method

SparseGaussianProcess(k::CovarianceKernel)

Creates a SparseGaussianProcess with kernel k, with 64 random features and 10 inducing points by default.

SparseGaussianProcesses.SquaredExponentialKernel — Type

SquaredExponentialKernel

A squared exponential kernel

$k(x,x') = \sigma^2\exp\left( -\left\lVert\frac{x - x'}{\kappa} \right\rVert^2 \right)$

parameterized by log-variance $\ln(\sigma^2)$ (trainable by default) and log-length-scales $\ln(\kappa)$ (trainable by default) applied element-wise to each dimension.

SparseGaussianProcesses.SquaredExponentialKernel — Method

(k::SquaredExponentialKernel)(x1::AbstractMatrix, x2::AbstractMatrix)

Computes the kernel matrix for the given squared exponential kernel.

SparseGaussianProcesses.SquaredExponentialKernel — Method

SquaredExponentialKernel(dim::Int)

Creates a squared exponential kernel of dimension dim.

Random.rand! — Method

rand!(self::EuclideanRandomFeatures, k::EuclideanKernel, 
      num_features::Int = size(self.frequency,ndims(self.frequency)))

Draw a new set of random features, by randomly sampling a new frequencies from the spectral measure, and new phases uniformly from $(0, 2\pi)$. Does NOT automatically resample the GP containing the features.

Random.rand! — Method

rand!(gp::GaussianProcess; num_samples = num_samples(gp))

Draw a new set of samples from the Gaussian process.

SparseGaussianProcesses.loss — Method

loss(gp::GaussianProcess, x::AbstractMatrix, y::AbstractMatrix; 
                                             n_data::Int = size(x,ndims(x)))

Computes the Kullback-Leibler divergence of the variational family from the posterior Gaussian process, up to an additive constant. Minimizing this function trains the Gaussian process.

SparseGaussianProcesses.CircularKernel — Type

CircularKernel

An abstract covariance kernel defined over a circle, or more generally a $d$-dimensional torus.

SparseGaussianProcesses.CovarianceKernel — Type

CovarianceKernel

A covariance kernel, assumed symmetric in its arguments.

SparseGaussianProcesses.CrossCovarianceKernel — Type

CrossCovarianceKernel

A cross-covariance kernel, NOT necessarily symmetric in its arguments.

SparseGaussianProcesses.EuclideanKernel — Type

EuclideanKernel

An abstract covariance kernel defined over a Euclidean space.

SparseGaussianProcesses.GaussianProcess — Type

GaussianProcess

An abstract Gaussian process type.

SparseGaussianProcesses.GaussianProcess — Method

(gp::GaussianProcess)(x::AbstractArray{<:Any,3})

Evaluate gp at a batched set of points x. Returns a 3-array whose dimensions are ordered (output_dimension, data_point_index, batch_index).

SparseGaussianProcesses.GaussianProcess — Method

(gp::GaussianProcess)(x::AbstractMatrix)

Evaluate gp at a set of points x. Returns a 3-array whose dimensions are ordered (output_dimension, data_point_index, sample_index).

SparseGaussianProcesses.Hyperprior — Type

Hyperprior

An abstract hyperprior.

SparseGaussianProcesses.InducingPoints — Type

InducingPoints

An abstract set of inducing points.

SparseGaussianProcesses.InterDomainOperator — Type

InterDomainOperator

An abstract inter-domain operator.

SparseGaussianProcesses.Kernel — Type

Kernel

An abstract kernel.

SparseGaussianProcesses.Kernel — Method

(k::Kernel)(x1::AbstractArray{<:Any,3}, x2::AbstractMatrix)

Computes a kernel matrix in batched form.

SparseGaussianProcesses.LeftGradientKernel — Method

(k::LeftGradientKernel{<:SquaredExponentialKernel})(x1::AbstractMatrix, 
                                                    x2::AbstractMatrix)

Computes the kernel matrix for the given gradient squared exponential cross-covariance.

SparseGaussianProcesses.PriorBasis — Type

PriorBasis

An abstract basis used for sampling a prior Gaussian process, with weights.

SparseGaussianProcesses.PseudoDataInducingPoints — Type

PseudoDataInducingPoints

A set of inducing points representing pseudo-data points with diagonal non-constant error covariance.

SparseGaussianProcesses.PseudoDataInducingPoints — Method

(self::PseudoDataInducingPoints)(z::AbstractMatrix, k::Kernel)

Sets the inducing locations of $u$ to $z$, inducing mean to zero, and inducing error variance to one for each pseudo-data point.

SparseGaussianProcesses.PseudoDataInducingPoints — Method

PseudoDataInducingPoints(k::Kernel, dim::Int, num_inducing::Int)

Creates a set of pseudo-data inducing points with unit error variance.

SparseGaussianProcesses.PseudoDataInducingPoints — Method

(self::PseudoDataInducingPoints)()

Assembles the pseudo-data inducing error variance into matrix form. Returns inducing locations, inducing mean, nothing jitter term, and inducing error variance.

SparseGaussianProcesses.RightGradientKernel — Method

(k::RightGradientKernel{<:SquaredExponentialKernel})(x1::AbstractMatrix, 
                                                     x2::AbstractMatrix)

Computes the kernel matrix for the given gradient squared exponential cross-covariance.

SparseGaussianProcesses.num_samples — Method

num_samples(gp::GaussianProcess)

Returns the number of random samples currently stored in gp.

SparseGaussianProcesses.pairwise_column_difference — Method

pairwise_column_difference(x::AbstractMatrix, y::AbstractMatrix)

Computes the 3-dimensional distance array, where out dimensions are (input_dimension, x_data_point_dimension, y_data_point_dimension).

SparseGaussianProcesses.prior_KL — Method

prior_KL(gp::SparseGaussianProcess{<:Any,<:Any,<:Any, <:Any,
                                   <:MarginalInducingPoints})

Computes the prior Kullback-Leibler divergence for a Gaussian process with marginal inducing points, given by the expression

$KL(q(u) \mathbin{||} p(u)) = \frac{1}{2} \left( \ln\frac{|K_{zz}|}{|\Sigma|} + tr(K_{zz}^{-1}\Sigma) + \mu^T K_{zz} \mu \right)$

where the mean is re-parameterized according to $\mu = \mathbb{E}( (K_{zz} + \xi I)^{-1} u )$.

SparseGaussianProcesses.spectral_distribution — Method

spectral_distribution(k::CircularSquaredExponentialKernel, 
                      num_samples::Integer)

Draws n samples from the spectral distribution of a standard squared exponential kernel on the circle.

SparseGaussianProcesses.spectral_distribution — Method

spectral_distribution(k::SquaredExponentialKernel; num_samples::Integer)

Draws n samples from the spectral distribution of a standard squared exponential kernel, which is multivariate Gaussian with covariance $2 I$.

Index

SparseGaussianProcesses.CircularKernel

SparseGaussianProcesses.CircularSquaredExponentialKernel

SparseGaussianProcesses.CircularSquaredExponentialKernel

SparseGaussianProcesses.CircularSquaredExponentialKernel

SparseGaussianProcesses.CovarianceKernel

SparseGaussianProcesses.CrossCovarianceKernel

SparseGaussianProcesses.EuclideanKernel

SparseGaussianProcesses.EuclideanRandomFeatures

SparseGaussianProcesses.EuclideanRandomFeatures

SparseGaussianProcesses.EuclideanRandomFeatures

SparseGaussianProcesses.EuclideanRandomFeatures

SparseGaussianProcesses.EuclideanRandomFeatures

SparseGaussianProcesses.GaussianProcess

SparseGaussianProcesses.GaussianProcess

SparseGaussianProcesses.GaussianProcess

SparseGaussianProcesses.GradientKernel

SparseGaussianProcesses.GradientOperator

SparseGaussianProcesses.Hyperprior

SparseGaussianProcesses.IdentityOperator

SparseGaussianProcesses.InducingPoints

SparseGaussianProcesses.InterDomainOperator

SparseGaussianProcesses.Kernel

SparseGaussianProcesses.Kernel

SparseGaussianProcesses.LeftGradientKernel

SparseGaussianProcesses.MarginalInducingPoints

SparseGaussianProcesses.MarginalInducingPoints

SparseGaussianProcesses.MarginalInducingPoints

SparseGaussianProcesses.MarginalInducingPoints

SparseGaussianProcesses.NormalHyperprior

SparseGaussianProcesses.NormalHyperprior

SparseGaussianProcesses.PriorBasis

SparseGaussianProcesses.PseudoDataInducingPoints

SparseGaussianProcesses.PseudoDataInducingPoints

SparseGaussianProcesses.PseudoDataInducingPoints

SparseGaussianProcesses.PseudoDataInducingPoints

SparseGaussianProcesses.RightGradientKernel

SparseGaussianProcesses.SparseGaussianProcess

SparseGaussianProcesses.SparseGaussianProcess

SparseGaussianProcesses.SquaredExponentialKernel

SparseGaussianProcesses.SquaredExponentialKernel

SparseGaussianProcesses.SquaredExponentialKernel

Random.rand!

Random.rand!

SparseGaussianProcesses.loss

SparseGaussianProcesses.num_samples

SparseGaussianProcesses.pairwise_column_difference

SparseGaussianProcesses.prior_KL

SparseGaussianProcesses.spectral_distribution

SparseGaussianProcesses.spectral_distribution